fix: refactor files to files folder
