+++ scopes/react/react/templates/react-component-js/files/component.ts
