+++ scopes/react/react/templates/react-component-js/files/composition.ts
