+++ scopes/react/react/templates/react-component-js/files/docs.ts
