+++ scopes/react/react/templates/react-component-js/files/test.ts
