+++ scopes/react/react/templates/react-component/files/component.ts
