+++ scopes/react/react/templates/react-component/files/composition.ts
