+++ scopes/react/react/templates/react-component/files/docs.ts
