+++ scopes/react/react/templates/react-component/files/test.ts
